Add tests and create `ParentProfileBackend` with business logic
Our coding challenge didn't have any tests in it, which I really want it to,
so that any potentiel candidates can showcase their testing skills.

I choose Scalatest as the test framework, as this is also what we're using
in our codebase.

In order to make the code unit-testable I had to change quite a few things;
	- Create `ParentProfileBackend` to contain the testable business logic
	- Move business logic from `ProfileRepository` to `ParentProfileBackend`
	- Change `GraphQLSchema` to use the pure `ParentProfileBackend` functions sandwiched between the impure `ProfileRepository`

Because `ProfileRepository` no longer contained any business logic, the function
names no longer made a lot of sense - the responsibility of `ProfileRepository`
is only about persisting changes to the database (CRUD), so I renamed the
functions to reflect that; creating, retrieving, updating and deleting.
The use-case specific names are now in `ParentProfileBackend`.

There is a gotcha, though. The domain model, in `Models.scala`, is also used
as rows in `ProfileRepository`. So they act as both the domain model and the
data model. This is not too big of a deal, but it does mean that the domain
model has to deal with `id` and know about incrementing it when creating new
ones.
A cleaner way would be to separate them, but I want to keep it as simple
as possible and it's not really a problem in a coding challenge like this.
In real life this wouldn't work for any distributed system as id's would
collide all the time - ideally the database should create the id itself, but
that means mapping back and forth between the same types in `Models`, which
I didn't find ideal.