Create tests for the nodejs backend too
Recently I added tests for the Scala backend. Here are the tests for the
nodejs backend also. I made them exactly the same as in Scala and also
tried to keep the implementation very similar.

I went with Jest as the testing framework, because that's also what we use
internally - in order to get it work with Docker I had to setup Jest with
Babel, which requires the various configuration files to work.

When running the backend through Docker, the tests will execute whenever
you change a file and the server will restart.